// src/store/redux/reducers/mines.type.ts
export enum EMinesSocketAction {
    CREATE_NEW_MINESGAME = '(MINES)-CREATE_NEW_MINESGAME',
    LOGIN_MINES = '(MINES)-LOGIN_MINES',
    SUBSCRIBE_MINES = '(MINES)-SUBSCRIBE_MINES',
    RECEIVE_ERROR = '(MINES)-RECEIVE_ERROR',
    DISCONNECT_MINES = '(MINES)-DISCONNECT_MINES',
    RESET_GAME_STATE = '(MINES)-RESET_GAME_STATE',
    MINES_ROLLING = '(MINES)-ROLLING',
    MINESGAME_ROLLED = '(MINES)-ROLLED',
    CASHOUT_MINESGAME = '(MINES)-CASHOUT_MINESGAME'
}